added ebnf grammar and operator precedence list
also whilst doing that I noticed some ambiguity in the grammar.
specifically in range constructs, fixing it with anew keyword should do
the trick. I plan on removing more stuff from the grammar, tho I'm not
sure as for what to remove.

The functional nature of how this goes restricts me quite a bit. I need
to study on how to overcome this restriction. I'll start writing the
parser tomorrow, or now idk. Hopefully, the language's parser will be
over by the week but I can't guarantee.

Gonna ask @isidentical a bit about this, but until then, stay home!