wrote half of the parser
This was BORING as HELL. Nevertheless I resisted the temptation of
leaving the project unfinished and made it this far. I'll write the rest
tomorrow.

Also, I added a REPL to the cli now, altought it shaked the other parts
a bit so that might segfault.. but who cares lol.

P.S.: The commit title looks so stupid :rofl: