removed premake
I think it's too much of an overkill.
Plus, now people don't have to go and download an executable just to
build another one.

So I count this as a win, though the lc module is incapable of compiling
currently. The problems began to resolve in my head slowly and I think I
can refactor this better now.

The Makefile is a mess, but it works.