:sparkles: added 4 basic arithmetic operators
they all work exclusively with integers for now with the only exception being
the modulus operator, which is yet to be implemented.

on top of being very naive, the way I model memory resources seems to be
rather unusable (thanks intel) with their current form. A more flexible
approach will be needed!